Update keybindings.md for clarity
New to Regolith and because this is not a nested list (like in the Helper) it was unclear that you needed to be in Resize Mode for `+ -` to change your gaps. I recently updated from 1.2 and believe there it was simply ⊞ Win-+/- so I assumed this was a bug until figuring out i needed to be in resize mode. 

Added "in Resize Mode" because this was the only line where that was not clear :)