Remove dollar signs in front of commands
When going to copy the installation commands for Regolith, I (and I assume most people) double/triple click to select the line and copy + paste. When doing this with the dollar signs in front of the commands it gives an invalid command. There is no need for the dollar signs here and it saves everyone a little bit of time and frustration by not having them.